[APP-FIX] 추천 산책 장소 마커 표시

+++ Client/app/src/main/java/com/ssafy/ccd/src/dto/Walk.kt
@@ -1,7 +1,7 @@
 package com.ssafy.ccd.src.dto
 
 data class Walk(
-    val lat: Int,
-    val lng: Int,
+    val lat: Double,
+    val lng: Double,
     val place: String
 )
